create SDCardManager for information about active file
