+++ redeem/SDCardManager.py
@@ -1,0 +1,11 @@
+from multiprocessing import Lock
+
+
+class SDCardManager(object):
+    current_file = None
+    current_line_count = None
+    current_file_count = None
+    current_lock = None
+
+    def __init__(self):
+        self.current_lock = Lock()
